Move database types to shared package
