+++ packages/teleport/src/Databases/ConnectDialog/ConnectDialog.tsx
@@ -22,10 +22,10 @@ import Dialog, {
   DialogContent,
   DialogFooter,
 } from 'design/Dialog';
+import { DbProtocol } from 'shared/services/databases';
 
 import { AuthType } from 'teleport/services/user';
 import TextSelectCopy from 'teleport/components/TextSelectCopy';
-import { DbProtocol } from 'teleport/services/databases';
 
 export default function ConnectDialog({
   username,
