+++ packages/teleport/src/services/databases/types.ts
@@ -14,6 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+import { DbProtocol } from 'shared/services/databases';
+
 import { AgentLabel } from 'teleport/services/agents';
 
 export interface Database {
@@ -25,14 +27,6 @@ export interface Database {
   names?: string[];
   users?: string[];
 }
-
-export type DbType = 'redshift' | 'rds' | 'gcp' | 'self-hosted';
-export type DbProtocol =
-  | 'postgres'
-  | 'mysql'
-  | 'mongodb'
-  | 'sqlserver'
-  | 'redis';
 
 export type DatabasesResponse = {
   databases: Database[];
